+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * #251, #345 Add ``bottom-left`` slot to table
 * #334 Add ``defaultDialogConfirmText`` and ``defaultDialogCancelText`` constructor options
 * #335 Add ``icon`` and ``iconPack`` prop to dialog
+* #350 Add Function type to ``component`` prop of Modal (thanks @ivalkenburg)
 
 ## 0.5.3